feat: add support for hotplugging and multiple Stream Decks
