+++ examples/fill-button-when-pressed.js
@@ -3,17 +3,17 @@
 const path = require('path');
 const streamDeck = require('../index');
 
-streamDeck.on('down', keyIndex => {
+streamDeck.on('down', (device, keyIndex) => {
 	// Fill the pressed key with an image of the GitHub logo.
-	streamDeck.fillImageFromFile(keyIndex, path.resolve(__dirname, 'github_logo.png'))
+	device.fillImageFromFile(keyIndex, path.resolve(__dirname, 'github_logo.png'))
 		.catch(error => {
 			console.error(error);
 		});
 });
 
-streamDeck.on('up', keyIndex => {
+streamDeck.on('up', (device, keyIndex) => {
 	// Clear the key when it is released.
-	streamDeck.clearKey(keyIndex);
+	device.clearKey(keyIndex);
 });
 
 streamDeck.on('error', error => {
